requirements.txt
fastapi
uvicorn[standard]
redis
pydantic
scikit-learn
pandas
joblib
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
psycopg2-binary==2.9.9